Discord: bugfix add/remove subs

+++ src/FplBot.Discord/Handlers/SlashCommands/RemoveSubscriptionsSlashCommandHandler.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Discord.Net.Endpoints.Hosting;
 using Discord.Net.Endpoints.Middleware;
 using FplBot.Discord.Data;
+using FplBot.Formatting;
 
 namespace FplBot.Discord.Handlers.SlashCommands
 {
@@ -25,46 +26,57 @@ namespace FplBot.Discord.Handlers.SlashCommands
         {
             var existingSub = await _repo.GetGuildSubscription(context.GuildId, context.ChannelId);
 
-            if (existingSub == null)
+            if (existingSub == null || !existingSub.Subscriptions.Any())
             {
-                return Respond($"⚠️ Did not find any sub in this channel to remove!");
+                return Respond($"🤷‍♀️O RLY?", $"Did not find any subscription(s) in this channel to remove!");
             }
 
             EventSubscription eventSub = Enum.Parse<EventSubscription>(context.CommandInput.Value);
 
-            bool existingIsAll = existingSub.Subscriptions.Count() == 1 && existingSub.Subscriptions.First() == EventSubscription.All;
-            bool hasAllAndUnsubAll = existingIsAll && eventSub == EventSubscription.All;
-
-            if (hasAllAndUnsubAll)
+            bool isLastSub = existingSub.Subscriptions.Count() == 1 && existingSub.Subscriptions.First() == eventSub;
+            if (existingSub.LeagueId == null && (isLastSub || eventSub == EventSubscription.All))
             {
                 await _repo.DeleteGuildSubscription(context.GuildId, context.ChannelId);
-                return Respond($"Unsubbed all events in this channel.");
+                return Respond($"✅ Success!", $"Removed subscription to this channel.");
             }
-
+            bool existingIsAll = existingSub.Subscriptions.Count() == 1 && existingSub.Subscriptions.First() == EventSubscription.All;
             if (existingIsAll && eventSub != EventSubscription.All)
             {
                 var allTypes = EventSubscriptionHelper.GetAllSubscriptionTypes().ToList();
                 allTypes.Remove(EventSubscription.All);
                 await _repo.UpdateGuildSubscription(existingSub with { Subscriptions = allTypes });
                 var updatedFromAll = await _repo.GetGuildSubscription(context.GuildId, context.ChannelId);
-                return Respond($"No longer subscribing to all events. Updated list: {string.Join(",", updatedFromAll.Subscriptions)}");
+                return Respond($"✅ Success!", $"No longer subscribing to all events. Updated list:\n{Formatter.BulletPoints(updatedFromAll.Subscriptions)}");
             }
 
-            var existingSubsWithout = new List<EventSubscription>(existingSub.Subscriptions);
-            existingSubsWithout.Remove(eventSub);
+            var updated = new List<EventSubscription>(existingSub.Subscriptions);
 
-            await _repo.UpdateGuildSubscription(existingSub with { Subscriptions = existingSubsWithout });
+            if (eventSub == EventSubscription.All)
+            {
+                updated = new List<EventSubscription>();
+            }
+            else
+            {
+                updated.Remove(eventSub);
+            }
+
+            await _repo.UpdateGuildSubscription(existingSub with { Subscriptions = updated });
             var regularUpdate = await _repo.GetGuildSubscription(context.GuildId, context.ChannelId);
-            return Respond($"Unsubscribed from {eventSub}. Updated list: {string.Join(",", regularUpdate.Subscriptions)}");
+            if (regularUpdate.Subscriptions.Any())
+            {
+                return Respond($"✅ Success!", $"Unsubscribed from {eventSub}. Updated list:\n{Formatter.BulletPoints(regularUpdate.Subscriptions)}");
+            }
+            return Respond($"✅ Success!", $"No longer subscribing to any events.");
+
         }
 
-        private static ChannelMessageWithSourceEmbedResponse Respond(string content)
+        private static ChannelMessageWithSourceEmbedResponse Respond(string title, string content)
         {
             return new ChannelMessageWithSourceEmbedResponse()
             {
                 Embeds = new List<RichEmbed>
                 {
-                    new($"ℹ️", content)
+                    new(title, content)
                 }
             };
         }
